+++ src/blocks/info-box/attributes.js
@@ -8,6 +8,10 @@ const attributes = {
 		selector: "span.uagb-ifb-title-prefix",
 		default: "Prefix",
 	},
+	classMigrate: {
+		type: "boolean",
+		default: false
+	},
 	infoBoxTitle: {
 		source: "html",
 		selector: "h1,h2,h3,h4,h5,h6",
